api test for statistics endpoint

+++ src/main/java/com/dmarkhas/n26/challenge/api/StatisticsDto.java
@@ -13,5 +13,5 @@ public class StatisticsDto {
 
     private double min;
 
-    private double count;
+    private long count;
 }
